Organize code block tests
